requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.2.5
idna==2.8
importlib-metadata==1.4.0
inflect==4.0.0
more-itertools==8.1.0
multidict==4.7.4
python-dateutil==2.8.1
six==1.14.0
websockets==8.1
yarl==1.4.2
zipp==2.0.0
